WIP: refactor CSS again
