+++ libgraffiti/src/css/values/mod.rs
@@ -3,4 +3,4 @@ mod color;
 mod dimension;
 mod enums;
 
-pub use {box_shadow::*, color::*, dimension::*, enums::*};
+pub use {box_shadow::BoxShadow, color::Color, dimension::Dimension, enums::*};
